+++ src/Client/ManagesCacheTrait.php
@@ -70,6 +70,10 @@ trait ManagesCacheTrait
         $this->cache->delete($prefix . ':browse');
         $this->cache->delete($prefix . ':browseAll');
         $this->cache->delete($prefix . ':writeType');
+
+        foreach ([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 14, 15, 16, 17, 18, 19, 20, 21, 22, 26] as $attr) {
+            $this->cache->delete($prefix . ':readMeta:' . $attr);
+        }
     }
 
     /**